+++ homeassistant/components/nuki/sensor.py
@@ -13,7 +13,7 @@ from homeassistant.const import PERCENTAGE, EntityCategory
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
-from . import NukiConfigEntry
+from .coordinator import NukiConfigEntry
 from .entity import NukiEntity
